🚚 rename: validationRules 를 common으로 옮기고 업데이트 #36
